services/retriever/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
sentence-transformers==3.0.1
transformers==4.41.2
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.26.3
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu

git+https://github.com/dorianbrown/rank_bm25.git@2550648efdbdcc5ebadbc8e5c8b26f5eb94b2b36#egg=rank_bm25
